HHB 3.1 Priest Release Candidate adjustments (#872)
* adjust margins to fix text shifts, fix missing comma in spell list

* remove changelog and add backpage for Priest, fix first-page layouting issues

* Oops, cleanup of content spill in Partials doc

* Updating Priest Class table to have Devotions in the table over a feature

* Updating Renew Faith to reduce space taken

* Removed Divine Versatility and fixed some Formatting from it

Removed Divine Versatility mostly due to it bordering on Tasha's content, plus with the shift to moving towards 2024 content, it bakes what this feature does into the feature itself, letting people more easily change some stuff out. So itd feel odd that priest be the only one getting something like this while others get updated with this in mind

* Added Unwavering Faith at 14th and fixed some formatting

Unwavering Faith is meant to mimic the 2024 Cleric where it improves the option given at 7th. Plus it felt odd that Only Divine Strikes got a boost at 14th level and not the other two options, so this also keeps it in line a bit.

* fix typo in Divine Word description

* tidy up textline alignment

* fix slight page size discrepancy

* don't show footer graphic on last page

* empowering faith should note that it is a 7th level feature

---------

Co-authored-by: OmNomDom308 <65504953+OmNomDom308@users.noreply.github.com>